fix: remove InvalidHealth error case in showAnswer server action
Previously, this server action only return success if the match's health
is enough (health is above 0 after showing answer).

Changed this action so that it allow showing answer for all health (even
if health is 0 or below after showing answer).

Don't need to worry about malicious action because when health <= 0, no
new question can be fetched or no new answer can be submitted.

Reason: so that I can show answer when the match end (health = 0).